chore(CorpusClient): Rename singleton for clarity
