+++ app/graphql/resolvers/corpus_slate_lineup_resolvers.py
@@ -52,7 +52,7 @@ async def resolve_home_slate_lineup(root, info: Info, locale: str = 'en-US') -> 
         default_value=DEFAULT_RECOMMENDATION_COUNT))
 
     slate_provider_kwargs = {
-        'corpus_fetchable': di.corpus_client,
+        'corpus_fetchable': di.corpus_feature_group_client,
         'recommendation_surface_id': RecommendationSurfaceId.HOME,
         'corpus_engagement_provider': di.corpus_engagement_provider,
         'locale': locale_model,
@@ -63,7 +63,7 @@ async def resolve_home_slate_lineup(root, info: Info, locale: str = 'en-US') -> 
         unleash_provider = UnleashProvider(graph_client_session, unleash_config=UnleashConfig())
 
         slate_lineup_model = await HomeDispatch(
-            corpus_client=di.corpus_client,
+            corpus_client=di.corpus_feature_group_client,
             preferences_provider=UserRecommendationPreferencesProvider(di.aioboto3_session, topic_provider),
             user_impression_cap_provider=di.user_impression_cap_provider,
             topic_provider=topic_provider,
