+++ tests/unit/data_providers/conftest.py
@@ -2,6 +2,6 @@ import pytest  # noqa
 
 from tests.mocks.corpus_engagement_provider import *  # noqa
 from tests.mocks.caching import *  # noqa
-from tests.mocks.corpus_feature_group_client import *  # noqa
+from tests.mocks.corpus_clients import *  # noqa
 from tests.mocks.translation_provider import translation_provider  # noqa
 from tests.mocks.topic_provider import topic_provider_en_us  # noqa
